// ha-wyoming-realtime-shim/src/types.ts
export interface WyomingMessage {
  type: string;
  data?: Record<string, unknown>;
}

export interface WyomingInfoMessage extends WyomingMessage {
  type: 'info';
  data: {
    stt: SttInfo[];
  };
}

export interface SttInfo {
  name: string;
  description: string;
  installed: boolean;
  languages: string[];
  version: string;
}